requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via
    #   email-validator
    #   pymongo
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via -r requirements.in
fastapi-cli==0.0.4
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.3
    # via fastapi
pydantic==2.7.2
    # via fastapi
pydantic-core==2.18.3
    # via pydantic
pygments==2.18.0
    # via rich
pymongo==4.7.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.30.1
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
